+++ examples/MNIST/MNIST-Cryptonets-Relu/common.py
