+++ examples/MNIST/MNIST-Cryptonets/W_squash.txt
