+++ examples/MNIST/pyclient_mnist.py
